telegram_bot/requirements.txt
python-telegram-bot
selenium
python-magic